Move `actionLinks` API
