--- apps/meteor/app/action-links/client/index.ts
@@ -1,3 +1,0 @@
-import { actionLinks } from './lib/actionLinks';
-
-export { actionLinks };
